[x2cpg] Clean-up and removal of deprecated API
Removed API does not have any downstream usage (codescience).
