--- joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/utils/StringUtils.scala
@@ -1,7 +1,0 @@
-package io.joern.x2cpg.utils
-
-implicit class StringUtils(str: String) {
-  def isAllUpperCase: Boolean = {
-    str.forall(c => c.isUpper || !c.isLetter)
-  }
-}
