--- joern-cli/frontends/x2cpg/src/test/scala/io/joern/x2cpg/utils/IgnoreInWindows.scala
@@ -1,10 +1,0 @@
-package io.joern.x2cpg.utils
-
-import org.scalatest.{Ignore, Tag}
-
-object IgnoreInWindows
-    extends Tag(if (System.getProperty("os.name").toLowerCase().contains("windows")) {
-      classOf[Ignore].getName
-    } else {
-      ""
-    })
